nixos: remove ranger and use index-fm instead
In CLI, zoxide and fzf work well, so the file manager is not necessary. The
reason why I need a file manager is that it can interact with other GUI
applications. Therefore, I choose one of GUI file manager instead of cli one.
Furthermore, I want a picture preview in filemanager, which is terriable in cli
and the thumbnails mode of feh is not worked greatly.